display logic

+++ modules/org.openhs.core.commons/src/org/openhs/core/commons/Window.java
@@ -6,8 +6,8 @@ public class Window extends Thing {
 	public String imagePath_close = "no path";
 	
 	//Coordinates in house 
-	public int x = 0;
-	public int y = 0;
-	public int z = 0;	
+	public float x = 0.0f;
+	public float y = 0.0f;
+	public float z = 0.0f;	
 
 }
